requirements.txt
# Web scraping
httpx>=0.27.0
playwright>=1.40.0
beautifulsoup4>=4.12.0
lxml>=5.0.0
selectolax>=0.3.21

# Data processing
pandas>=2.1.0
pydantic>=2.5.0

# Async support
anyio>=4.0.0

# Rate limiting and retries
tenacity>=8.2.0
ratelimit>=2.2.1

# Database
sqlalchemy>=2.0.0
aiosqlite>=0.19.0

# Scheduling
schedule>=1.2.0
apscheduler>=3.10.0

# Utilities
python-dotenv>=1.0.0
rich>=13.7.0
loguru>=0.7.0

# Export formats
openpyxl>=3.1.0
